Tong

+++ src/main/java/com/example/techshop/entity/OrderDetailEntity.java
@@ -17,10 +17,10 @@ public class OrderDetailEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_detail_id")
-    private int orderDetailsId;
+    private Integer orderDetailsId;
 
     @Column(name = "total")
-    private Long total;
+    private Integer total;
 
     @Column(name = "ispaid")
     private Boolean ispaid;
